Pull out a generic type for state-based strategy pattern to ensure type safety+coherence across the UIBox class
