+++ Assets/Scripts/UI/UIBox/UIBoxBase.cs.meta
@@ -1,0 +1,3 @@
+fileFormatVersion: 2
+guid: 45c1c38b129a4f20bcd2d6a12854c8e1
+timeCreated: 1784898523
